+++ src/Commands/SystemCommands/GenericCommand.php
@@ -11,7 +11,6 @@
 namespace Longman\TelegramBot\Commands\SystemCommands;
 
 use Longman\TelegramBot\Commands\SystemCommand;
-use Longman\TelegramBot\Entities\ServerResponse;
 
 /**
  * Generic command
@@ -31,21 +30,5 @@ class GenericCommand extends SystemCommand
     /**
      * @var string
      */
-    protected $version = '1.0.0';
-
-    /**
-     * Command execute method
-     *
-     * @return ServerResponse
-     */
-    public function execute()
-    {
-        //$message = $this->getMessage();
-        //$chat_id = $message->getChat()->getId();
-        //$user_id = $message->getFrom()->getId();
-        //$command = $message->getCommand();
-        //$text = trim($message->getText(true));
-
-        return parent::execute();
-    }
+    protected $version = '1.1.0';
 }